1. 增加了facer和game两个子项目
2. jersey项目中增加权限管理模块

+++ jersey/src/main/java/org/ranji/lemon/jersey/controller/UserController.java
@@ -60,8 +60,8 @@ public class UserController {
 	
 	@RequestMapping(value="/delete",method =RequestMethod.POST)
 	@ResponseBody
-	public String delete(@RequestParam(value = "ids[]")  Integer[]  ids){
-		List<Integer> list= Arrays.asList(ids); 
+	public String delete(@RequestParam(value = "ids[]")  Long[]  ids){
+		List<Long> list= Arrays.asList(ids); 
 		
 		userService.deleteByIDS(list);
